+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactor `OverdueListDownloader`
 - Bump flipper to v0.153.0
 - Set target aspect ration and rotation for QR code scanner preview
+- Use `OverdueListDownloader` for generating overdue patients CSV or PDF using `OverdueCsvGenerator`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes